next part of common.inc.php cleanup
