+++ viewlogs.php
@@ -3,6 +3,7 @@
 use Utils\Database\OcDb;
 use Utils\Uri\Uri;
 use lib\Controllers\LogEnteryController;
+use Utils\Text\TextConverter;
 
 
 
@@ -305,7 +306,7 @@ if ($error == false) {
 
             if ($record['text_html'] == 0) {
                 $processed_text = htmlspecialchars($processed_text, ENT_COMPAT, 'UTF-8');
-                $processed_text = help_addHyperlinkToURL($processed_text);
+                $processed_text = TextConverter::addHyperlinkToURL($processed_text);
             } else {
                 $processed_text = userInputFilter::purifyHtmlStringAndDecodeHtmlSpecialChars($processed_text);
             }
